sqlite_adapter: restructure tests and add sqlite test module
